fix(worktrees): collect ignored nested checkouts
